Merged with what I forgot to push before. Removed extraneous type definition. Defined the dataUnion in the header as public. This is because we cannot return a pointer to a private array. Also updated the return data function to return a pointer (because we want an array, not just a char).
Also put two unions in for the collection time (public) and in a previous commit added a private union for the index. To find the time since the last collection, I added a new function that merely returns the time from the private array (i.e. point.timeSince()). this is to help keep our class from having too many public variables and behaving in unexpected ways.