updated serialtest to not send newline or carriage return chars unless explicitly defined at the top. This is to attempt to resolve issues with the communicating with the radios, where an escape char was successfully sent but threw it out of diagnostic mode. I was suspicious that it was the newline that was being sent afterward, so the code was rewritten to explicitly say whether a newline would be sent or not.
Probably, the newline thing can be easily worked into the program's arguements; at the moment, however, it should not be necessary (no one really wants unwanted newlines, anyway). However, they way it is written, a newline cannot even be escaped because the write implementation will not send the newlines wherever they appear. Maybe this can be resolved in a different way (like loading the entire stdin buffer, removing the trailing newlines and carriages, and then writing that). However, as this program was originally intended to merely test the abilities of c++ to communicate over serial, it is unlikely that this will get fixed.